Add files via upload
Update: The sqlite database now includes a sample of 12 HSK 2.0 level 1 words, sorted by ascending alphabetical order of their pinyin.  The sample also includes 17 sentences that each incorporate at least one of the 12 sample words, but are strictly limited to vocabulary with the HSK 2.0 level 1 word list.  With the exception of adding more sentence IDs to the word table as new sentences within HSK 2.0 level 1 are added, this sample is in its final form.  This is to say that all future additions to these tables will follow the same style guidelines as the sample submission unless a change to the formatting style is made later.  In this case, the sample and all future additions will be altered to match the new style.