Update: add database samples in JSON form
Attatched is the sample of the HSK 2.0 Level 1 vocabulary.  The format is two separate JSON files: one for the word table and one for the sentence table.  Once the python script is debugged, the script will be used to append the JSON files to their respective tables in an SQLite database.  The final version will be a SQLite database of the same format.  All JSON files and SQLite databases will follow the same style guide.  Because the python script will only allow for appending to the database and not editing, the entries of the database will be added one HSK level at a time.