Update: word table data for HSK 1 minus sentences
Attached is the data for the HSK 2.0 Level 1 vocabulary.  The data is in JSON form and will need to be appended to the SQLite database via the python script once finalized.  All of the words are included, but the Level 1+ words will be updated as sentences are added to the sentence table.  Level 1+ words are words that are not included in HSK 2.0, but if they were, they would be around a Level 1, and are used by practice sentences for Level 1.  Adding the Level 1+ vocabulary will allow the app to define all words used in the example sentences later down the road, while also letting app users get accustomed with ambiguity and needing to rely on context clues to  understand new vocabulary.  Additionally, the sentence attribute has not been completed, but will be updated after the sentence table data has been finalized.  Because the python script can only append, the JSON data should only be converted and appended after all changes are made to both this file and the sentence data file.