Update: add HSK 2.0 Level 1 sentence data
Attached is the data for the sentence table of the HSK 2.0 Level 1 vocabulary and practice sentence database.  The file is in JSON form and is not completed.  Sentences have been added for the first 20% of the vocabulary list.  More practice sentences will be uploaded soon.  The JSON file should eventually be appended to the SQLite database in the sentence table, but only after all of the data is finalized, since the script does not allow for editing, only for appending.